requirements.txt
django>=4.2.11,<5.0
pillow
django-modeltranslation
django-crispy-forms
crispy-bootstrap5
django-ckeditor
